// src/services/suparAdmin/type.ts
export interface School {
    id?: string;
    schoolName: string;
    schoolAddress: string;
    city: string;
    state: string;
    schoolLandlineNo: string;
    ownerName: string;
    gst: string;
    boardType: string;
    adminContact: string;
    serviceStartDate: string;
    currentPlan: string;
    subscriptionType: string;
    renewalDate: string;
    status: string;
    email: string;
    password: string;
    landline: string;
}